shiva-railway/requirements.txt
metaapi-cloud-sdk
pandas
pandas_ta
python-dotenv
numpy
